fix tests

+++ queue/test/fiskaltrust.Middleware.Localization.QueueDE.IntegrationTest/SignProcessorDETests/Services/TarFileCleanupTests.cs
@@ -22,7 +22,7 @@ using Moq;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace fiskaltrust.Middleware.Localization.QueueDE.IntegrationTest.SignProcessorDETests.Closings
+namespace fiskaltrust.Middleware.Localization.QueueDE.IntegrationTest.SignProcessorDETests.TarFileCleanup
 {
     public class TarFileCleanupTests : IClassFixture<SignProcessorDependenciesFixture>
     {
@@ -34,7 +34,67 @@ namespace fiskaltrust.Middleware.Localization.QueueDE.IntegrationTest.SignProces
         }
 
         [Fact]
-        public async Task SignProcessor_DailyClosingReceipt_ShouldFinishAllOpenTransactions()
+        public async Task SignProcessor_TarFileCleanup_ShouldDeleteTarFiles()
+        {
+
+            var receiptRequest = JsonConvert.DeserializeObject<ReceiptRequest>(File.ReadAllText(Path.Combine("Data", "DailyClosingReceipt", "Request.json")));
+            var expectedResponse = JsonConvert.DeserializeObject<ReceiptResponse>(File.ReadAllText(Path.Combine("Data", "DailyClosingReceipt", "Response.json")));
+            var queueItem = new ftQueueItem
+            {
+                cbReceiptMoment = receiptRequest.cbReceiptMoment,
+                cbReceiptReference = receiptRequest.cbReceiptReference,
+                cbTerminalID = receiptRequest.cbTerminalID,
+                country = "DE",
+                ftQueueId = Guid.Parse(receiptRequest.ftQueueID),
+                ftQueueItemId = Guid.Parse(expectedResponse.ftQueueItemID),
+                ftQueueRow = expectedResponse.ftQueueRow,
+                request = JsonConvert.SerializeObject(receiptRequest),
+                requestHash = "test request hash"
+            };
+            var queue = new ftQueue
+            {
+                ftQueueId = Guid.Parse(receiptRequest.ftQueueID),
+                StartMoment = DateTime.UtcNow
+            };
+
+            await AddOpenOrdersAsync();
+
+            var journalRepositoryMock = new Mock<InMemoryJournalDERepository>()
+            {
+                CallBase = true
+            };
+            journalRepositoryMock.Setup(x => x.InsertAsync(It.IsAny<ftJournalDE>())).CallBase();
+            var actionJournalRepositoryMock = new Mock<IActionJournalRepository>(MockBehavior.Strict);
+            var config = new MiddlewareConfiguration { Configuration = new Dictionary<string, object>() { { Constants.ConfigurationKeys.STORE_TEMPORARY_FILES_KEY, false } }, QueueId = queue.ftQueueId, ServiceFolder = Directory.GetCurrentDirectory() };
+            var configurationRepository = _fixture.CreateConfigurationRepository();
+
+            if (Directory.Exists(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
+            {
+                foreach (var f in Directory.GetDirectories(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
+                {
+                    Directory.Delete(f);
+                }
+
+                foreach (var f in Directory.GetFiles(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
+                {
+                    File.Delete(f);
+                }
+            }
+
+            var tarFileCleanupService = new TarFileCleanupService(Mock.Of<ILogger<TarFileCleanupService>>(), journalRepositoryMock.Object, config);
+            var sut = RequestCommandFactoryHelper.ConstructSignProcessor(Mock.Of<ILogger<SignProcessorDE>>(), configurationRepository, journalRepositoryMock.Object, actionJournalRepositoryMock.Object,
+                _fixture.DeSSCDProvider, new DSFinVKTransactionPayloadFactory(), new InMemoryFailedFinishTransactionRepository(), new InMemoryFailedStartTransactionRepository(),
+                _fixture.openTransactionRepository, Mock.Of<IMasterDataService>(), config, new InMemoryQueueItemRepository(), new SignatureFactoryDE(config), tarFileCleanupService);
+
+
+            var (receiptResponse, actionJournals) = await sut.ProcessAsync(receiptRequest, queue, queueItem);
+
+            Directory.GetFiles(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")).Should().HaveCount(0);
+            Directory.GetDirectories(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")).Should().HaveCount(0);
+        }
+
+        [Fact]
+        public async Task SignProcessor_TarFileCleanup_ShouldDeleteOldTarFiles()
         {
 
             var receiptRequest = JsonConvert.DeserializeObject<ReceiptRequest>(File.ReadAllText(Path.Combine("Data", "DailyClosingReceipt", "Request.json")));
@@ -68,16 +128,18 @@ namespace fiskaltrust.Middleware.Localization.QueueDE.IntegrationTest.SignProces
             var config = new MiddlewareConfiguration { Configuration = new Dictionary<string, object>() { { Constants.ConfigurationKeys.STORE_TEMPORARY_FILES_KEY, true } }, QueueId = queue.ftQueueId, ServiceFolder = Directory.GetCurrentDirectory() };
             var configurationRepository = _fixture.CreateConfigurationRepository();
 
-            foreach (var f in Directory.GetDirectories(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
+            if (Directory.Exists(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
             {
-                Directory.Delete(f);
-            }
+                foreach (var f in Directory.GetDirectories(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
+                {
+                    Directory.Delete(f);
+                }
 
-            foreach (var f in Directory.GetFiles(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
-            {
-                File.Delete(f);
+                foreach (var f in Directory.GetFiles(Path.Combine(config.ServiceFolder, "Exports", config.QueueId.ToString(), "TAR")))
+                {
+                    File.Delete(f);
+                }
             }
-
 
             var tarFileCleanupService = new TarFileCleanupService(Mock.Of<ILogger<TarFileCleanupService>>(), journalRepositoryMock.Object, config);
             var sut = RequestCommandFactoryHelper.ConstructSignProcessor(Mock.Of<ILogger<SignProcessorDE>>(), configurationRepository, journalRepositoryMock.Object, actionJournalRepositoryMock.Object,
